Update Rename-TabletopTokenImage.ps1
Directory for unmatched tokens should only be created if a match wasn't found. This provides an indication to the user that manual action is required.

An empty directory for unmatched tokens adds visual clutter to the output directory, and more importantly introduces the potential for accidentally copying it to the production tokens directory.